+++ crates/affinidi-tdk/common/affinidi-tdk-common/src/config.rs
@@ -220,7 +220,7 @@ impl TDKConfigBuilder {
     }
 
     #[cfg(feature = "messaging")]
-    /// How many Authentication sets should we cache?
+    /// Should TDK create an ATM instance internally?
     /// Defaults: true
     /// Example:
     /// ```